moved and updated upload controller
