+++ boat-bay-server/src/main/java/com/backbase/oss/boat/bay/web/views/dashboard/controller/BoatDocumentationController.java
@@ -14,11 +14,12 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 import org.springframework.web.servlet.ModelAndView;
 
 
-@Controller
+@RestController
 @RequestMapping("/api/boat/")
 @Transactional
 @RequiredArgsConstructor
